Included google analytics
Included google analytics:
-Time spent on analytics page
-Time spent on application
